switch lib to queries/mutations/actions
